Removing LoaderImpl and refacotr to have only one Loader that uses the TaskManager in engine
