+++ grakn-engine/src/main/java/ai/grakn/engine/controller/ImportController.java
@@ -19,8 +19,6 @@
 package ai.grakn.engine.controller;
 
 import ai.grakn.engine.loader.Loader;
-import ai.grakn.engine.loader.LoaderImpl;
-import ai.grakn.engine.loader.client.LoaderClient;
 import ai.grakn.engine.postprocessing.PostProcessing;
 import ai.grakn.exception.GraknEngineServerException;
 import ai.grakn.graql.Graql;
@@ -44,7 +42,6 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
 import java.util.concurrent.Executors;
@@ -52,14 +49,12 @@ import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import static ai.grakn.engine.controller.Utilities.getAsList;
 import static ai.grakn.engine.controller.Utilities.getAsString;
 import static ai.grakn.engine.controller.Utilities.getKeyspace;
 import static spark.Spark.before;
 import static spark.Spark.halt;
 import static spark.Spark.post;
 import static ai.grakn.util.REST.Request.PATH_FIELD;
-import static ai.grakn.util.REST.Request.HOSTS_FIELD;
 import static java.util.concurrent.TimeUnit.SECONDS;
 
 @Api(value = "/import", description = "Endpoints to import Graql data from a file.")
@@ -90,20 +85,18 @@ public class ImportController {
                     "be used to load ontologies because it splits up the file into smaller parts.")
     @ApiImplicitParams({
             @ApiImplicitParam(name = "path", value = "File path on the server.", required = true, dataType = "string", paramType = "body"),
-            @ApiImplicitParam(name = "hosts", value = "Collection of hosts' addresses.", required = true, dataType = "string", paramType = "body"),
             @ApiImplicitParam(name = "keyspace", value = "Name of graph to use", dataType = "string", paramType = "query")
     })
     private String importDataREST(Request req, Response res) {
         try {
             final String keyspace = getKeyspace(req);
             final String pathToFile = getAsString(PATH_FIELD, req.body());
-            final Collection<String> hosts = getAsList(HOSTS_FIELD, req.body());
 
             final File file = new File(pathToFile);
             if (!file.exists())
                 throw new FileNotFoundException(ErrorMessage.NO_GRAQL_FILE.getMessage(pathToFile));
 
-            Loader loader = getLoader(keyspace, hosts);
+            Loader loader = getLoader(keyspace);
 
             // Spawn threads to load and check status of loader
             ScheduledFuture scheduledFuture = scheduledPrinting(loader);
@@ -121,11 +114,10 @@ public class ImportController {
     /**
      * Return the appropriate loader- Blocking if no hosts are provided, distributed otherwise
      * @param keyspace name of the graph to use
-     * @param hosts collection of hosts' addressed
      * @return Loader configured to the provided keyspace
      */
-    private Loader getLoader(String keyspace, Collection<String> hosts){
-        return hosts == null ? new LoaderImpl(keyspace) :new LoaderClient(keyspace, hosts);
+    private Loader getLoader(String keyspace){
+        return new Loader(keyspace);
     }
 
     /**
@@ -135,7 +127,7 @@ public class ImportController {
      */
     private ScheduledFuture scheduledPrinting(Loader loader){
         return Executors.newSingleThreadScheduledExecutor()
-                .scheduleAtFixedRate(() -> LOG.info(loader.getLoaderState()), 10, 10, SECONDS);
+                .scheduleAtFixedRate(loader::printLoaderState, 10, 10, SECONDS);
     }
 
     // This method works under the following assumption:
