+++ grakn-migration/base/src/main/java/ai/grakn/migration/base/io/MigrationCLI.java
@@ -22,9 +22,6 @@ import ai.grakn.Grakn;
 import ai.grakn.GraknGraph;
 import ai.grakn.engine.GraknEngineServer;
 import ai.grakn.engine.backgroundtasks.standalone.StandaloneTaskManager;
-import ai.grakn.engine.loader.Loader;
-import ai.grakn.engine.loader.LoaderImpl;
-import ai.grakn.engine.loader.client.LoaderClient;
 import ai.grakn.util.Schema;
 import com.google.common.io.Files;
 import ai.grakn.graql.InsertQuery;
