+++ grakn-migration/export/src/main/java/ai/grakn/migration/export/GraphWriter.java
@@ -39,6 +39,7 @@ public class GraphWriter {
 
     public GraphWriter(GraknGraph graph){
         this.graph = graph;
+        this.graph.showImplicitConcepts(true);
     }
 
     /**
